feat: 빈 오브젝트의 역할과 구분
